added week comparison, favorites, avoid
cleaned up some wording, added a tool to show the best day of the week, if you just want to go at a certain time. added the ability to avoid and favor lanes. changed the players and games to be input validated. added a feature that users' last 5 searched addresses will remember in the box.